code moi nhat

+++ DA_QuanLyCuaHangBanDT/src/main/java/com/poly/entity/ChiTietDonHang.java
@@ -28,7 +28,7 @@ public class ChiTietDonHang {
     private Integer soLuong;
 
     @Column(name="Gia")
-    private Float gia;
+    private Double gia;
     
     @ManyToOne
     @JoinColumn(name = "Madh")
